requirements.txt
backports.functools-lru-cache==2.0.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.3.1
cycler==0.12.1
fonttools==4.53.1
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
oauth2client==4.1.3
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
subprocess32==3.5.4
tqdm==4.66.4
uritemplate==4.1.1
urllib3==2.2.2
google-auth-oauthlib==1.2.1